+++ incubator/binding-risingwave/src/main/java/io/aklivity/zilla/runtime/binding/risingwave/internal/statement/RisingwaveCreateMaterializedViewTemplate.java
@@ -14,25 +14,60 @@
  */
 package io.aklivity.zilla.runtime.binding.risingwave.internal.statement;
 
-import net.sf.jsqlparser.statement.Statement;
+import net.sf.jsqlparser.statement.create.table.CreateTable;
 import net.sf.jsqlparser.statement.create.view.CreateView;
 
 public class RisingwaveCreateMaterializedViewTemplate extends RisingwaveCommandTemplate
 {
     private final String sqlFormat = """
         CREATE MATERIALIZED VIEW IF NOT EXISTS %s AS %s;\u0000""";
+    private final String fieldFormat = "%s, ";
+    private final String includeFormat = "COALESCE(%s, zilla_%s_header::varchar) as %s, ";
 
     public RisingwaveCreateMaterializedViewTemplate()
     {
     }
 
     public String generate(
-        Statement statement)
+        CreateView createView)
     {
-        CreateView createView = (CreateView) statement;
         String view = createView.getView().getName();
         String select = createView.getSelect().toString();
 
         return String.format(sqlFormat, view, select);
+    }
+
+    public String generate(
+        RisingwaveCreateTableCommand command)
+    {
+        CreateTable createTable = command.createTable;
+        String name = createTable.getTable().getName();
+
+        String select = "*";
+
+        if (command.includes != null)
+        {
+            fieldBuilder.setLength(0);
+
+            createTable.getColumnDefinitions()
+                .forEach(c -> fieldBuilder.append(
+                    String.format(fieldFormat, c.getColumnName())));
+            command.includes.forEach((k, v) ->
+            {
+                if ("timestamp".equals(k))
+                {
+                    fieldBuilder.append(String.format(fieldFormat, v));
+                }
+                else
+                {
+                    fieldBuilder.append(String.format(includeFormat, v, v, v));
+                }
+            });
+
+            fieldBuilder.delete(fieldBuilder.length() - 2, fieldBuilder.length());
+            select = fieldBuilder.toString();
+        }
+
+        return String.format(sqlFormat, "%s_view".formatted(name), "SELECT %s FROM %s_source".formatted(select, name));
     }
 }