+++ runtime/engine/src/main/java/io/aklivity/zilla/runtime/engine/catalog/CatalogHandler.java
@@ -69,6 +69,12 @@ public interface CatalogHandler
         return NO_VERSION_ID;
     }
 
+    default int[] unregister(
+        String subject)
+    {
+        return new int[0];
+    }
+
     String resolve(
         int schemaId);